[general] Wait on semaphore in main loop

+++ src/zjs_modules.h
@@ -17,18 +17,13 @@
  *
  * @param handle        Handle that was registered
  *
- * @return              1 if the routine did any processing
- *                      0 if the routine did not process anything
- *
- * Note: The return value is only used for the auto-exit Linux feature. If
- *       0 is returned (and there are no timers or callbacks) AND auto-exit
- *       is enabled, it will cause the program to exit.
+ * @return              Time until next event, ZJS_TICKS_FOREVER if unknown
  */
-typedef uint8_t (*zjs_service_routine)(void *handle);
+typedef int32_t (*zjs_service_routine)(void* handle);
 
 void zjs_modules_init();
 void zjs_modules_cleanup();
-void zjs_register_service_routine(void *handle, zjs_service_routine func);
-uint8_t zjs_service_routines(void);
+void zjs_register_service_routine(void* handle, zjs_service_routine func);
+int32_t zjs_service_routines(void);
 
 #endif  // __zjs_modules_h__
